requirements-build.txt
# Build-time dependencies (pip install -r requirements-build.txt)
# Use with:  pip install -r requirements.txt -r requirements-build.txt
# Then run:  python build.py

pyinstaller>=6.0.0
Pillow>=10.0.0
